// src/app/Models/Checkout/Checkout-Model.ts
export class Checkout {
    uniqueOrderNumber?: string;
    trackingID?: string;
    first_name?: string;
    last_name?: string;
    email_address?: string;
    mobile?: string;
    company_name?: string;
    address_line_one?: string;
    address_line_two?: string;
    city?: string;
    postcode?: string;
    country?: string;
    regionOrState?: string;
    orderNotes?: string;
    orderDate?: string;
    // expectedDelivery?: string;
    // orderTime?: string;
    orderStatus?: string;
    paymentType?: string;
    creditCardType?: string;
}